Make build scripts slightly better

+++ build_linux.sh
@@ -1,20 +1,32 @@
 #!/bin/bash
+
+clean_build_dirs () {
+    [ -d ./__pycache__ ] && rm -rf __pycache__
+    [ -d ./build ] && rm -rf ./build
+    [ -d ./venv ] && rm -rf ./venv
+    [ -f ./onthespot_linux.spec ] && rm ./onthespot_linux.spec
+    [ -f ./onthespot_linux_ffm.spec ] && rm ./onthespot_linux_ffm.spec
+}
+
+
 echo "========= OnTheSpot Linux Build Script ==========="
 echo " => Cleaning up !"
-rm -rf __pycache__
-if [ -f "./dist/onthespot_linux" ]; then
-    rm ./dist/onthespot_linux
-fi
-rm -rf ./build
-rm -rf ./venv
+[ -f ./dist/onthespot_linux ] && rm ./dist/onthespot_linux
+[ -f ./dist/onthespot_linux_ffm ] && rm ./dist/onthespot_linux_ffm
+clean_build_dirs
+
 echo " => Creating virtual env."
 python3 -m venv venv
+
 echo " => Switching to virtual env."
 source ./venv/bin/activate
+
 echo " => Installing 'pyinstaller' via pip..."
 pip install pyinstaller
-echo " => Installing dependencies pip..."
+
+echo " => Installing dependencies to venv with pip..."
 pip install -r requirements.txt
+
 if [ -f "ffbin_nix/ffmpeg" ]; then
     echo " => Found 'ffbin_win' directory and ffmpeg binary.. Using ffmpeg binary append mode "
     pyinstaller --onefile \
@@ -35,11 +47,11 @@ else
                 --icon="src/onthespot/resources/icon.png" \
                 src/portable.py
 fi
-echo " => Setting permissions.. "
-chmod +x ./dist/onthespot_linux
-echo " => Cleaning.. "
-rm onthespot_linux.spec
-rm onthespot_linux_ffm.spec
-rm -rf ./build
-rm -rf __pycache__
+echo " => Setting executable permissions.. "
+[ -f ./dist/onthespot_linux ] && chmod +x ./dist/onthespot_linux &>./build_nix.log
+[ -f ./dist/onthespot_linux_ffm ] && chmod +x ./dist/onthespot_linux_ffm &>./build_nix.log
+
+echo " => Cleaning .. "
+clean_build_dirs
+
 echo " => Done "
